Corrects key size for 3DES and AES256 Ciphers
When the assignment was origianlly submitted the cipher suites listed
incorrect key sizes. 3DES is 192bits: 3 sets * 64bits. 8 bits of each 64
key is reserved. This should yield a 168bit key. However, due to the
meet in the middle attack only 112bits are actually effective.

AES256, as the name suggests uses 256bits or 32 bytes for encryption.